chore(useUpdateNode): small refactor
